Rename _VisuallyHidden -> _ToggleVisibility to prevent confusion
